implement rest api

+++ courses/src/main/java/ru/tembaster/courses/service/ProfessorService.java
@@ -21,12 +21,17 @@ public class ProfessorService {
         return professorRepository.save(professor);
     }
 
-    public void delete(Professor professor) {
-        professorRepository.delete(professor);
+    public boolean delete(int id) {
+        Professor professor = professorRepository.findById(id).orElse(null);
+        if (professor != null) {
+            professorRepository.delete(professor);
+            return true;
+        }
+        return false;
     }
 
     public Professor get(int id) {
-        return professorRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid id: " + id));
+        return professorRepository.findById(id).orElse(null);
     }
 
     public List<Professor> getAll() {
